Create 815. Bus Routes.cpp
This approach will build up the graph using the bus stops as the nodes. To connect the edges in this graph, we need to find the stop we can go from a particular stop. There can be multiple routes that have this stop, we can go to any bus stop that is present in all these routes. One way is to store all these bus stops in the routes that have this stop, but this would take up too much memory. Instead, we can only store the indices of the routes that have the stop. The next reachable stop from this stop will be all the bus stops on all these routes.

So, we have prepared a graph where the nodes are bus stops and we can find the next stop we can travel to from each stop. We need to find the shortest distance between two given nodes "source" and "target". Since the edges are unweighted, BFS is more appropriate than Dijkstra's algorithm.

In the problem statement, it's given that we are not on any bus initially. Hence, to start from the bus stop source we can board any of the bus that has the source as one of the stops in its route. So the breadth-first search here needs to be a multi-source BFS starting with all the buses that have the source in stops. During the BFS, we will pop the first bus from the queue and iterate over the stops that this route involves. For each stop, we will check if this stop is equal to target, and then we can return the current count of buses busCount. If the stop is not equal to target then we will iterate over all the routes that have this stop and add them to the queue if the route is not visited before. Note that we are keeping track of visited routes instead of bus stops because when we visit a route we are essentially visiting all the stops in that route and hence keeping track of visited stops individually is not that efficient.

If we have completed the BFS and still haven't reached the target, it implies there is no way to reach that stop and hence we can return -1.

Algorithm

Return 0 if the source and target are the same.

Initialize an empty map from an integer to a list of integers adjList to store the edges. The key is the bus stop and the value is the list of integers denoting the indices of routes that have this stop.

Initialize an empty queue q and an unordered set vis to keep track of visited routes.

Insert the initial routes into the queue q and mark them visited in vis.

Iterate over the queue while it's not empty and do the following:

Pop the route from the queue.
Iterate over the stops in the route.
If the stop is equal to target, return busCount.
Otherwise, iterate over the routes for this stop in the map adjList.
Add the unvisited routes to the queue and mark them visited.
Return -1 after completing the BFS.