Create 2785. Sort Vowels in a String.cpp
Intuition

Given a string s having English lowercase or uppercase letters, we need to build a new string such that the vowels in the original strings are arranged in the non-decreasing order of their ASCII values.

We don't want to break the original order of the consonants, so we will not sort the whole string. Instead, we need to sort only the vowels. Therefore, we will collect all the vowels from s into a string temp, and sort it in ascending order. Now, we need to put these characters back into the original string s in the sorted order, that is, the first vowel in the original string s will be replaced with the first character in temp, the second vowel will be replaced with the second character in temp, and so on. This way, the consonants will remain in the original order and the vowels will be sorted in the ascending order.

In some languages where strings are immutable, we can use other mutable data structures provided by that language to collect and sort all the vowel characters.

Algorithm

Create a method isVowel that returns true if the given character is a lowercase or uppercase vowel, and returns false otherwise.
Iterate over the string s and store the vowels in the string temp.
Sort the string temp in ascending order.
Initialize an empty string ans to store the answer string, and an integer j to 0 to track the current index in the string temp.
Iterate over the string s and for each character, if the character is a vowel, we add the character temp[j] to ans and increment j, otherwise, we add the character from s.
Return ans.