Create 1846. Maximum Element After Decreasing and Rearranging.cpp
In this problem, we need to maximize any value under the following rules:

arr[0] = 1.
Each adjacent element differs by at most 1.
We can decrease any value, but we can't increase values.
We can rearrange values.
For an array of length n, the biggest value we could have is n. This scenario would be when the array is [1, 2, 3, ..., n].
This is because each adjacent element differs by at most 1, so the best we could do is to just count up from 1, which the first element must be. When is this scenario impossible?

Because we are not allowed to increase elements, the best-case scenario is impossible when the original elements are not large enough to support the counting.
So what should our strategy be? We start at the first index with our answer ans = 1. This is because by the rules of the problem, the first element must be equal to 1. Now, we iterate over the rest of the indices and try to increment by 1 each time.

If we can successfully increment, we update ans = ans + 1. We can successfully increment if there is an element in arr that is greater than or equal to ans + 1. Any element that is greater than or equal to ans + 1 can be reduced to ans + 1 according to the rules. However, once we reduce it, we can't use that element anymore in the future.

So which element should we choose at each step? We should greedily choose the smallest element that is greater than or equal to ans + 1. The reason we want the smallest element is because choosing a larger element does not give us any additional benefit - we will only increment our answer by 1 regardless. However, choosing the smallest element "saves" the larger elements to be reduced in the future.

For example, let's say you had ans = 3 and there was a 4 and a 5 in the array. If you chose to reduce the 5 to a 4, you would not be able to reach ans = 5 anymore. However, if we use the 4 instead, then the 5 remains available when we want to increment ans to 5.

Note that because we are allowed to rearrange elements freely, their initial order is irrelevant. As such, we will start by sorting arr so we can process the elements in ascending order.

We also initialize ans = 1 and begin iterating over arr, starting from index 1. The reason we skip index 0 is because arr[0] = 1 - we have no choice. At each index i, we try to increment ans by using arr[i]. If arr[i] is greater than or equal to ans + 1, then we can reduce arr[i] (or keep it the same) to ans + 1.

Algorithm

Sort arr in ascending order.
Initialize ans = 1.
Iterate i over the indices of arr, starting from i = 1:
If arr[i] >= ans + 1, increment ans.
Return ans.