Create 1980. Find Unique Binary String.py
There are many more possible binary strings than there are strings in nums.

In fact, since there are only n strings in nums, we never need to check more than n + 1 different binary strings, since at least one of them would not appear in nums and thus be a valid answer. How do we decide which n + 1 binary strings we should check?

Let's start by converting each string in nums to its equivalent base-10 integer. We will store these integers in a hash set integers. Now, we can simply use a for loop to iterate over the range [0, n] (the size of this range is n + 1, so it is guaranteed to contain at least one valid answer). For each number, we check if it is in integers. If it isn't, it represents a valid answer. We just need to convert it back to a binary string of length n and return it.

Note that in some cases, if a valid answer, when converted to a binary string, has a length shorter than n, we need to add "0"s to the beginning to make its length equal to n.

Algorithm

Create integers, a hash set containing all the elements of nums in their base-10 integer form.
Initialize n = nums.length.
Iterate num from 0 to n:
If num is not in integers, convert it to a binary string of length n and return it.
The code should never reach this point. Return anything.