Create 1877. Minimize Maximum Pair Sum in Array.cpp
We are given an array of NNN integers (where NNN is even), and we need to pair up all these integers in such a way that the maximum sum of a pair is minimized. We need to return the minimized maximum pair sum in the array.

One might think of pairing up the smaller integers together so that the pair sum is minimized at the beginning. However, in this way, we will end up pairing the bigger integers together and hence will increase the maximum pair sum. For example, we can pair up the array [1, 1, 2, 3] as (1, 1), (2, 3) with maximum sum as 5. However, the optimal way would be to make a pair like (1, 3), (1, 2) with a maximum sum of 4. This suggests that pairing adjacent numbers by value may not be the optimal approach.

Observing the above example we can think about another potential strategy to pair up the smallest integer with the greatest integer and then second-smallest with the second-greatest and so on. This method ensures we don't end up pairing the two greatest integers as we did in the previous method. But how can we know this is optimal and would always produce the minimum pair sum?

Suppose we have the array $[a_{1}, a_1, a_2, ......, a_{n}]$, where these integers are sorted in ascending order and so a1a_{1}a 
1
​
  is the minimum integer and ana_{n}a 
n
​
  is the greatest one in the array. As per the above potential solution, we will pair up a1a_{1}a 
1
​
  and ana_{n}a 
n
​
  together, we can try to prove this method wrong by contradiction. Let's say we assume there exist two integers in the array $a_{i}$ and $a_{j}$ and satisfy this:

a1≤ai≤ana_{1} \le a_{i} \le a_{n}a 
1
​
 ≤a 
i
​
 ≤a 
n
​
 
a1≤aj≤ana_{1} \le a_{j} \le a_{n}a 
1
​
 ≤a 
j
​
 ≤a 
n
​
 
Let's assume an opposite method, and prove that it leads to contradiction. Suppose the pair$[(a_{1}, a_{i}), (a_{n}, a_{j}) ]$ is optimal than pair $[(a_{1}, a_{n}), (a_{i}, a_{j}) ]$. This is however not true, because $a_{j} + a_{n}$ is always bigger than or equal to $a_{i} + a_{1}$, hence the max of $[(a_{1}, a_{i}), (a_{n}, a_{j}) ]$ will always be $a_{j} + a_{n}$. And no matter what's the max of $[(a_{1}, a_{n}), (a_{i}, a_{j}) ]$ it will always be smaller than or equal to $a_{j} + a_{n}$. Therefore, our potential solution strategy is optimal.

The image below demonstrates the pairing:

fig

Therefore, we will sort the integers in the array and then pair the integers at the left end with the integers at the right end. The minimum value is paired with the maximum value, the next smallest value is paired with the next largest value, and so on. We need to iterate over only the first half of the array because the corresponding second element in the pair can be found by using the length of the array.

Algorithm

Sort the array nums.
Initialize the variable maxSum to 0.
Iterate over the array nums from index 0 to nums.length() / 2 - 1.
Get the sum of the current element and its corresponding pair nums[i] + nums[nums.length() - 1 - i], and update maxSum if the sum is larger.
Return maxSum.