Create 445. Add Two Numbers II.cpp
Our task is to do a basic addition of two numbers starting with the least significant digits and working our way up to the most significant digits. In the previous approach, we reversed the linked lists to access the least significant digits first. We can also use stacks to access the least significant digits first.

The advantage of using a stack is that when we loop over a given linked list from the first node to the last and push all the digits in the stack, the top of the stack will have the least significant digit and the bottom will contain the most significant digit.

We can add the digits at corresponding places of the linked lists using the two stacks moving from the least to the most significant digits using the stack's pop method.

Here's a brief visual representation explaining the approach:

img

Algorithm
Create two integer stacks s1 and s2 to store the integers of the linked lists l1 and l2 respectively.
Push all the integers of l1 in s1 starting from the integer at the first node. The most significant comes first in the list, so it will be stored at the bottom of the stack and the least significant digit will stored at the top.
Similarly, push all the integers of l2 in s2.
Create two integers totalSum and carry to store the sum and carry of current digits.
Create a new ListNode, ans that will store the answer.
We will add the two numbers present in the linked list now by adding the digits one by one. We continue until both s1 and s2 are empty:
If s1 is not empty, pop the first element from the stack and add it to totalSum.
If s2 is not empty, pop the first element from the stack and add it to totalSum.
Set ans.val = totalSum % 10.
Store the carry as totalSum / 10.
Create a new ListNode, newNode that will have val as carry. Set next of newNode to ans. Update ans = newNode to use the same variable ans for the next iteration.
Update totalSum = carry.
If carry == 0, it means the newNode that we created in the final iteration of while loop has val = 0. Because we perform ans = newNode at the end of each while loop, to avoid returning a linked list with a head of 0 (leading zero), we return the next element, i.e., we return ans.next. Otherwise, if carry is not equal to 0, the value of ans is non-zero. Hence, we just return ans.