Create 136. Single Number.cpp
The question simply asks us to find an element in the given array whose frequency is 1.All the other elements have a frequency=2.
We have to do so in :

Linear Time
Using Constant Space

Algorithm 

Sort the array.
Traverse the array and check if one of the adjacent elements is equal to the current element or not.
If yes , move ahead. Else return the current element.