Create 2210. Count Hills and Valleys in an Array.cpp
We make an array of integers v that contains the unique values of nums. We parse i from the second element until the second to last one and compare if v[i] is smaller or bigger than it's neighbours and increment the count.
I skipped the first and the last numbers because they dont have 2 neighbouring values.