Create 1424. Diagonal Traverse II.cpp
Let's think about the grid as a graph. Each square is a node, and we can imagine each node having an edge to the squares below and to the right (if they exist). Let's take a look at the diagonal image again:

img


As you can see, a node with identifier x has edges to nodes with identifier x + 1. If we consider the top-left square 0, 0 as a "source" node, then each square's identifier is exactly equal to its distance from the source. This allows us to visit the diagonals in order using BFS!

If you are not familiar with BFS, please check out the relevant LeetCode Explore Card.

We start a BFS from 0, 0. At each node row, col, we first push row + 1, col to the queue and then row, col + 1. Note that we only add a square to the queue if it both exists and has not been visited yet.

How do we know if a square has been visited yet? We could use a hash set to keep track of visited squares, but there is a simpler way. We only need to consider the square row + 1, col (down) if we are at the start of a diagonal. Otherwise, for every other square on the diagonal, the square below it has already been visited by the right edge of the previous square.

img


The level-wise nature of BFS will ensure that we visit all squares in a diagonal with identifier x before we visit any square in a diagonal with identifier x + 1. This means we will visit the diagonals in the correct order. Because we add the square row + 1, col before row, col + 1, we will also traverse over each diagonal in the correct order as well. This means our entire BFS will traverse the input in the same order as the answer, allowing us to solve the problem in one pass!

Algorithm

Initialize a queue with (0, 0) and the answer list ans.
While queue is not empty:
Remove (row, col) from queue.
Add nums[row][col] to ans.
If col == 0 and row + 1 is in bounds, add (row + 1, col) to queue.
If col + 1 is in bounds for the current row, add (row, col + 1) to queue.
Return ans.