Create 1630. Arithmetic Subarrays.cpp
The first thing to notice is that any arithmetic sequence must be sorted. This is because each successive element differs by a constant amount, so the entire sequence must be monotone since the change is constant.

For a given subarray arr, how do we check if we can form an arithmetic sequence? The problem states that we are allowed to rearrange arr. Thus, we should start by sorting arr, since if we can form an arithmetic sequence, the sequence must be sorted.

Once we have sorted arr, we can simply iterate over each adjacent element and check if the differences are constant. We will initialize diff = arr[1] - arr[0] as the difference between the first two elements.

img


We will then iterate over all other adjacent elements and check if their difference is equal to diff. If any difference is not equal to diff, then we cannot form an arithmetic sequence.

img


If all differences are equal to diff, then we can form an arithmetic sequence. This brings us to our solution. We will define a function check(arr) that takes a subarray arr and applies the above process to determine if it is an arithmetic sequence.

Then, we will iterate over all pairs l[i], r[i] and form arr as the subarray of nums from l[i] r[i]. Once we have arr, we will pass it into check to find the answer for the ithi^{th}i 
th
  query.

Algorithm

Define check(arr):
Sort arr.
Initialize diff = arr[1] - arr[0].
Iterate i over the indices of arr, starting from 2:
If arr[i] - arr[i - 1] != diff, return false.
Return true.
Initialize the answer ans.
Iterate i over the indices of l:
Create arr as the subarray of nums from indices l[i] to r[i].
Add check(arr) to ans.
Return ans.